requirements.txt
coverage==4.5.4
Django==2.2.3
django-crispy-forms==1.7.2
gunicorn==19.9.0
psycopg2-binary==2.8.3
PyJWT==1.7.1
pytz==2019.1
sqlparse==0.3.0
